feat: Action Queue auto dependency keys
Allow actions to specify dependency keys that any ohter action can
depend on without having to remember the previous action id. See
the documentation of `ActionDependencyKey` and
`ActionDependencyKesy` for more info.

This will allow us to enable parallel execution of actions in the
default group as we correctly setup scenarios such as:

* Ensuring Star/Unstar depends on the previous star operation
* Ensuring all action run after the last queued event loop action
* Ensuring label operation on a new label are auto reverted if the
  create label action fails.