feat(ET-1410): Sender Address Change
Allow the user to change the sender address of their draft.

Changing the sender address will cause the attachment key packets to
change and will need to be updated locally after the draft has
been saved on the server.

If we have a public key attachment setting enabled, previous public keys
that match the previous address used by this account will be removed
and replaced with the public key for the current address.

Finally, if the previous address signature is not empty and remains
intact, we will replace it with the new address signature.