refactor*: Remove rowid and weird proc macro hack and on_save, on_load fix.
The rowid field was useless but it had permeated the whole codebase, this commit removes it.

This also cleans up a bit how things were done in the orm, including parts of the on_save hack.

From now on, new overrides to on_save and on_load should use the
do_on_load and do_on_save shims in the trait.

This can be incrementally done, so it's not done here.