fix(ET-3169): Replace path separators from attachment file name
It appears that some users have attachment file names with path
separators in them. Usually this occurs when the user uploads
attachments for an OS with different path separation conventions.

Before storing the attachment we replace the local path separator with
'_' ensuring the file can be correctly created in the cache.