chore: Fix changelog generator
It was previously using an OrderedDict to hold the BFS job queue:

```
jobs.update({p: tags.get(p) or t for p in c.parents})
```

If a particular commit had multiple tags associated with it (caused by
merges and weird history), all but one of the tags would be lost;
the `jobs` remembered only the last tag of each commit.

This would have been a good performance optimization if we didn't care
which *which* tag a commit belonged to, but that's not the case here.
Switching to a list of tuples fixes the issue, at the expense of a
small increase in the runtime.