feat*: core: Implement fake clock
This commit implements a clock that can be asked about the current time,
and whose current time can be faked (set to a constant).

This is needed for RSVP tests (see the next commit), because answering
an RSVP sends a request that contains a timestamp inside of it (`{
"UpdateTime": 123123123, ... }`), so having a way of faking time comes
handy for asserting those types of requests.

Alternatively, the RSVP test could use a partial-request mock (i.e. a
mock that ignores the timestamped-field), but I think having the ability
to fake time will come handy later anyway, so why not implement it now.