fix: Incorrect DB Migration
Correct the v24 migration and run it again as v25 to resolve the
problems created by the incorrect v24 version.

The delete from query failed as it does work correctly when there are
null values present. We have to make sure that no null values
are returned in the select from draft_metadata table.