fix(ET-3422): Improve HTML address signature update
To ensure we can always update the signature when in HTML mode, wrap
the signature with a custom div and then replace the content of that
div with the new signature. This is similar to what web is doing and
should work correctly for all cases.

This patch also extends HTML Transformer with a `replace_inner_div`
method to perform the html replace.