fix: Make sure our tracing logger is not crashing the app if fails
Currently we are using .init() which panics whenever the logger cannot
be initialized. This might be bad, because we should **never** crash
the app.

I decided to just ignore the error instead.

Reasons:

* We might want to run `create_mail_session` multiple of times in tests,
so the second initialization of logger should fail silently
* Even if logger fails to initialize on the production, user still will
be able to use the app. It makes sense to silence the error instead of
showing them a popup "sorry, could not open proton mail because of
logger" or something else.