fix*: rsvp: Support cancelled events (better)
Cancelled events contain an `invite.ics` with `METHOD:CANCEL` - we need
to recognize it the same way we recognize `METHOD:REQUEST`, as otherwise
this bit returns `Err(IcsIsNotRsvpRequest)`:

https://gitlab.protontech.ch/rust/proton-rust/-/blob/dbfb7673679d91b7edda74f3bee2a04fcd5ed733/calendar/calendar-common/src/rsvp.rs#L66

... which is then caught here:

https://gitlab.protontech.ch/rust/proton-rust/-/blob/dbfb7673679d91b7edda74f3bee2a04fcd5ed733/mail/mail-common/src/mailbox/decrypted_message.rs#L425

... causing attachment not to be recognized as an RSVP.

Note that we already support cancelled events _per se_:

https://gitlab.protontech.ch/rust/proton-rust/-/blob/dbfb7673679d91b7edda74f3bee2a04fcd5ed733/calendar/calendar-common/src/rsvp/fetch.rs#L153

... as in: we already read the event status from the API - we don't need
to actively react to `METHOD:CANCEL` and cancel the event manually, no.

We just need to be able to extract the event id from within `invite.ics`
if that invite happens to be a cancellation.

(so in practice the current logic works for Google Calendar invites, as
those tend to be `RsvpEventId::Direct` thanks to headers injected by
backend, but we don't recognize cancellations for internal invites which
in practice are usually `RsvpEventId::Indirect`.)