feat(ET-3025): Add size based log rotator
Adds the log service crate which contains a log rotator that splits
the log files into up to N multiple chunks at a configurable size
limit.

The service also exposes method to export all logs into a single file
so it can be more easily accessed by other utilities.

The log service exists so that we can later move the entire
initialization of the log an other subscribers into this one crate.