fix: iOS 0xdead10cc
After the background worker completes its tasks in the allotted time,
we notify iOS via a callback that work has been completed.

After the callback triggers, iOS then queried `all_messages_sent` to
check if there was a user context which unsent messages.

If the conditions were right, this could trigger the creation of user
contexts outside of the permitted time window, which would trigger db
operations and eventually be terminated by the OS.

We have removed 2 methods which are considered "Dangerous" and now we
return the data required in the callback directly.