feat(ET-3677): Rework label_as and add the undo action to its return type
This adds the capability to undo (online and offline) the label_as
action, exposed in both uniffi and the tui.

The previous approach was too complex and was not really undoable,
and also by memoizing stuff it actually made things slower :derp:

It also hardcoded the archive into it, leading to a bug where rollback
would not unarchive it.
This simplifies it and fixes the bug by queuing the archive as another action.

I've also parallelized all network requests to try to make it faster.

I've also extracted common logic into a trait

This also adds the capability of also archiving to the TUI