feat*: rsvp: Warn about outdated invites
Each time organizer modifies an event in a meaningful way (e.g. changes
event's start/end date), attendees get a new `invite.ics` - opening an
older invite must then warn the user that they are looking at an
outdated invite, not the current one.

Note that as-is this change here is kinda meaningless - that's because
we always load data from the API, so the user can't actually see stale
data on the widget - and this, as it turns out, is wrong. Widget should
pull data mostly from the invite itself and only supplement it with the
API stuff in case something's missing.

(you can confirm this on web - if you invite yourself to an event and
then change the event's summary, opening the previous invite will show
you the _older_ summary as well; and our TUI app will show you the
current summary.)

But that warrants a separate pull request.