fix(ET-3759): Unable to reply to messages
When the handling of email alias was introduced, we used the value
of `X-Original-To` as is. This value often contains the email address
lower cased.

If the user had a mixed case address (e.g.: FooBar@proton.me) this would
set the sender with the lower cased version and the backend is unable to
correctly resolve this.

If an alias is present, we extract it and apply it to the user's
address.