fix(ET-3783): Missing attachments
While I was not able to reproduce the error in the ticket, i did find
some issues that could potentially be related.

The local_id of the attachment metadata on both Conversations and
Messages was not saved after write, which can lead to being unable to
open the data.

Scroller Eq for Conversation and Messages also did not take into account
changes to the attachment metadata. The former can change when new
messages arrive in a thread and the latter when updating drafts.

Finally, we also add some more logs to catch issues with sync/decrypting
attachment data.