fix: Improve scroller offline mode capabilities
The problem was initially sparked by muon not returning timeout errors in the observer's layer. This have caused Scroller to always try load even if the device was clearly offline. Additionally I've spotted opportunity to speed up loading not visited locations in the offline mode which actually made whole code less complex in the end - 2 birbs with one stone.