fix*: ical: Support quoted escapes
Trying to parse a quoted escape (e.g. `"hello \\ world"`) as a
`ParamValue` would cause the parser to enter an infinite loop due to an
accidentally-unimplemented code path.

Note that *.ics files don't support escaped strings anyway, so this is a
weird example to begin with, but it's rather easy to handle, so why not.

Found by fuzzing.