fix: Marking conversation unread without messages
There is a special path when we try to mark the conversation as unread
without having all the message metadata. Before stash the
conversation labels had different queries to update them. With stash,
the conversation labels are updated via the Conversation type. This in
turn would cause the `conversation.save()` to write out the previous
values again.