feat: Introduce AppDetails structure
# Warning: Breaking changes

This commit introduces `AppDetails` present in the uniffi layer.
It (in most cases) replaces `ApiConfig::app_version` field.

It requires an additioal work by the mobile developers, but the fix is
relatively trivial.

# Before

Before, we were asking the mobile side to provide App Version in the
format of: `{platform}-{product}@{version}` (example: "ios-mail@7.0.0").
We used that string to inform the API from where session comes from.

# After

Instead of concatenating on their own, now mobile developers need to fill
all three fields separaterly:
* platform
* product
* version

and it is Rust responsibility to concatenate it.

Upsides are:
* If the format changes, only Rust needs to be updated
* In share extension we can reuse the same information
  (platform & product) to fork the API session (`@version` is skipped)