refactor(ET-3484): Remember alias address when changing addresses
When we receive an email on an alias we need to remember it so the
user can change back to this alias when they change addresses.

The alias address will only be remembered when crafting a reply or if
it is present in the sender fields of the draft when we open it. This
matches the behavior on web.