feat: Parallel Default Action Groups
Finally enable parallel execution of the default action group.

Prefetching and rollbacks have been moved to their own queues so
they don't interfere with execution of the pending actions in the
default group.

Later we will also want to disable prefetching rollback in the
background execution contexts to save bandwidth.