feat(ET-4084): Draft recipient expiration feature check
We need to check if all recipients have password protection to display
a dialog to the user that only proton recipients support this feature.

To correctly support offline mode, we can only return information for
recipients we have already validated with a network request.

If the recipient is not yet validated or any of the other states it is
labeled as unknown.

Finally if password protection is enabled, this is always supported.