refactor(ET-3948)*: Draft Actor
Refactor the existing draft code into an actor system. This is 1:1
feature compatible port. There are things that can be improved and will
be handled in follow ups.

The reason we are doing this is due to logic that is contained outside
the mail-common crate that needs to be moved back
(e.g.: recipient validation) and we want to make sure that we have auto
save working correctly across all platforms.

This also solve a long standing problem with the TUI where we needed to
use special queuers for changes to be modified.

The next set of changes will focus on bringing the validation back into
the actor and providing an event stream for UI clients to update.