fix: [ET-4002] Support multipart emails with local-only attachments
Validation inside `generate_mime_top_package()` is too restrictive - the
code _used to_ require for all attachments to be uploaded up-front, but
that was overhauled when we were implementing RSVPs.

This code path is exercised when you enable the `Sign external messages`
option and use `PGP/MIME` scheme - of course, I've made sure that we
can, in fact, correctly generate an RSVP response in this case.