fix(ET-4244): Handle NeedNewPass state
Accounts with temporary passwords (e.g. private subusers during initial
login) need to set a password before they are able to be used; when they
set their password, we generate all their PGP keys.

However, if the user kills the app before setting a password, we would
leave the account in a weird state: the absence of a saved key secret in
the db would lead us to think the account is in the 'NeedMbp' state, but
that's not correct.

We need to track separately whether the account has a temporary password
that needs to be changed.