fix: Validate queue action state
Validate that all actions in the queue can be loaded and accessed before
starting the executors.

There are more elegant solutions to this problem, but at least this way
the clients can show an error message to log out and log back in again
rather than have some silent background failure.