refactor*: Manually remember last draft save action
We used to rely on the database to remember the last draft save action.
There is a small chance this could potentially fail since the value was
read out side of transaction scope, which could lead to out of date
values.

Since we now have an actor, we can remember the value ourselves and have
it always be correct. We still use the db value on draft open in case
there are pending save actions.