fix: vcard: Support empty `;TYPE=`
It seems that some vcards have emails with empty types:

```
ITEM2.EMAIL;TYPE=;PREF=2:foo.bar@example.com
           ^----^
```

I haven't been able to create such a vcard using Proton Contacts, so my
guess is that those contacts must've been imported from another system,
one that doesn't strictly conform to the spec.

That's because while the `TYPE` property is optional, if it's empty then
it shouldn't be printed whatsoever, there should be no `;TYPE=` string
in the input vcard in that case.

As it stands, our parser understands that as `types: [""]` (i.e. one
type that's an empty string) and fails to parse it.

This commit slightly adjusts the logic so that we skip over empty types,
in this case correctly parsing the file as `types: []` (which is
alright, since the type is an optional property anyway, so empty type =
missing type = gud).