fix(ET-4426): Do not decorate images and links by default
This matches the behavior of web's HTML to text conversion.
Unfortunately there is no way to force the text renderer to easily
insert a new line for the image element.

In the interest of time we manually remove a custom tag in this case. To
be checked post GA for better options.