fix: Excessive Ping
We have observed a problem that when the dns query fails it causes an
endless loop of retries. The retry policy of the request is not
respected. To address this we manually do the retry timeouts outside of
the network layer to prevent this .

`StatusWatcher` was also updated to not ignore the current update as
this can cause it to miss changes.