fix: Bug in which snooze_time was eqaul across pages.
This was problematic as in we had no snooze time and time fields which caused random items to be picked as end of the page. This caused very inefficient requests to the backend which in turn caused loading pages with 3/4 items in it. This looked like an infinite scroller loading in accounts which were snoozing by pages. The solution for that was to introduce both time fields in the cursor. This allowed to pick correct items for API request. Additionally the request now use anchoring instead of end/endID. This in theory allows API to consistently order items later converted to the cursors.