fix: Endless event loop enqueues
There were 2 bugs that caused an endless amount of event polls to be
queued, whether online or offline.

The first one was forcing the event loop loop action to be queued rather
than using the replace mechanism.

The second was related to a bug in the action queue where upon replacing
an action we did not remove the previous dependency keys for the
replaced actions, which would trigger a cyclic dependency error,
resulting in new event poll actions every time.