fix: Do not ping when os tells us we are offline
When the OS reports to us that we have no network connection we now stop
the background ping task.

As soon as the OS returns back online we always do a check to validate
this as this can be a false positive with VPNs. If this still fails we
restart the background ping.