fix(ET-4125)*: Ensure we don't infinitely trigger live queries
Since we changed the code to refetch the conversation from the api, we
introduced an infinite loop due to the conversation update re-triggering
the callback, which then cause the clients to call the same method
again.

We have to convert the current conversation to API representation in
order to check if they are different due the all the local data we add
onto it that is not available in a 1:1 conversion.