refactor(ET-4125): Open conversation from push notification
Sadly the previous approach, while fixing the problem lead to
undesirable outcomes for general user experience.

Since we only need this new behavior to handle opening push
notifications, we have separate this into a separate method and restored
the old method as the default.

When watching a conversation we now specify what the origin of the
request is and we act accordingly.

Previous methods for syncing /not syncing the conversation have been
removed.