fix: Action queue lost contexts
It is possible that under certain circumstances the queue can have
handlers registered that belong to a MailUserContext that has gone
out of scope.

To prevent dangling `LostContext` errors, we re-register all actions
which depend on a context.

We also make sure to wait for the next action if we detect a lost
context as the issue should solve it self in a short period of time.