fix(all): Remove redundant `.gitignore` files.
The auto-generated local `.gitignore` files overlapped with what was
defined in the `.gitignore` file for the monorepo. So we can remove
most of them.

I first wanted to remove the lines with `Cargo.lock` as well, but I
left this for the teams responsible for these projects.

The official stance on whether to version `Cargo.lock` or not has
changed over the years:
https://blog.rust-lang.org/2023/08/29/committing-lockfiles/.

In: the past, `cargo new` would add `Cargo.lock` to the `.gitignore`
file. The reasoning was that this would help ensure that people would
test new minor versions quickly and thus ensure better stability in
the ecosystem.

Later, `cargo new` stopped adding `Cargo.lock` to the ignore file.
This ensures better reproducibility since everyone who compiles from
source will use the exact same versions of all dependencies. As a
company with many developers working on the same codebase, I will urge
us to follow this practice to get stable and reproducible builds.