refactor: Change Event Poll queueing
Previously we always inserted an event poll after 30 seconds regardless
of what is in the queue.

We now change the code to queue after any action that is in the queue.
Rather than depending on the event loop, action simply record the event
loop key.

Next, the event loop action is replaced every time we queue a new
action in the default group so that it always runs after everything
has been completed.

Finally, forcing pull to refresh will now run without any dependencies
to act immediately.