refactor: Sync previous page is no longer influencing `fetch_more` directly.
The issue which have come with fetch_new is that we were able to provide multiple sync previous calls and then they would have been consumed in the fetch_more. Which would be correct and is but I noticed that this case was interpreded as Network error which is confusing clients so the load happens much slower. With this fix we provide data first with Append update and then Refresh comes with the new data as follow up.

This also simplifies the scroller flow to be independent of the hidden internal communication which was hard to follow and understand.