fix(ET-4011): Message Too Large
In this saga there was still a corner case error that can only be
detected at send time that does not fail during draft save.

If you forward/reply to a message which has a total attachment count
larger than 25mb, this will trigger any errors on save, but it will fail
at send time.

Other errors of this nature are tied to attachment uploads, which never
happen in this scenario. So, when we detect this, we return this new
error code.