feat(ET-4721): Issue Reporter Service
Introduce the proton-issue-report-service to manage the reporting of
issues to a monitoring platform such as sentry.

While we currently don't have a pure rust implementation we will use
this to hook up to the mobile client's implementation.

The global `IssueReporter` is present on `Context`. Each user will be
assigned their own `IssueReporter` via a
`IssueReporter::new_user_reporter`.

Finally, issues reported by the user should respect the telemetry
setting.