fix(ET-4806): Remove tracing-oslog until crash is fixed
Tracing-oslog crashes whenever it cannot reach span data.
Span data might be missing when the app goes into the background.

Instead of using `.expect()` tracing-oslog should handle things
gracefully. But it is not. Since it causes a crash, it is better to
remove that dependency for now (its not used in production build
anyway), and later re-introduce 0.3.1.