refactor: Fetch and apply events 1 by 1
We have observed that under certain circumstances event grouping can
include up to 20 events, which can lead to large delay before the user
sees changes as we need to make 20 separate network requests in those
cases.

Now we fetch the events 1 by 1 and apply them as they arrive so the
feedback to the user will be more immediate, even if the changes are not
yet complete.