fix(ET-4266): Proper validation of expiring message support
Previous logic relied on a value in the response that actually does not
signal that this is proton powered account, but an official Proton
account for communication.

We now check whether the email actually has a Proton powered encryption
key. This method is not 100% accurate, but according to the crypto team
it is the best way to retrieve this information.