fix(ET-4384): Only allow one event loop action to be queued
We have observed that clients/users can abuse the pull to refresh
action. One user had 100+ event polls queued up that were delaying
the processing of other actions.

We now cancel the actions so that they are always inserted at the end of
the queue rather than inheriting the current spot. Even though,
technically, the priority system should prevent this, it is simpler to
identify and reason about it. It also sidesteps potential dependency
issues with replacing actions.