fix(ET-4384): Also update messages when opening converastion
When opening a conversation from a push notification, we also need
to update the messages that come with it to make sure that label context
info matches with the messages.

A bug was reported where due to not syncing the new message state the
user could not see that one of the N messages was still unread and was
the reason for their conversation not transitioning to read state after
opening the new message.