refactor: Subscribing to tables is now async
Based on recent reports it's possible that the watcher's command queue
may be full and the sending of the command is blocked.

To avoid blocking the async executor, use `tokio::spawn_blokcing` to
perform the registration.

Stash workers are not affected since they are running in a synchronous
environment already.