fix: Force event poll
First, the forced to event poll was not executing the correct setup for
the `EventPoll` action we forgot to disable the dependency keys for this
scenario. Even though it was queued with a higher priority, it would
still wait on other actions in the queue to finish executing.

Second, periodic event poll had the potential to cancel higher priority
event poll if, for some reason it was not running yet. We now keep
separate action id states for normal and forced event polls to avoid
them cancelling each other out.