fix: Resync conversation if message count does not match
With recent changes to the scrollers to fetch new data when we pull
to refresh or the app returns to the foreground, it's possible for the
scroller to fetch new data before the event loop kicks in.

Under certain circumstances this can cause the UI to display
conversations with new messages, but when opening the conversation the
messages are not there yet.

We provide a quick fix to this problem where if we detect the local
number of messages does not match the conversation total, we resync the
conversation as if it were opened from a push notification.

Properly addressing this will require a bit more time and some api
changes and will be considered for future work.