feat(ET-5033): Support recipients with display names
Sometimes when you copy an email address from somewhere - shout-out to
SimpleLogin - you get a display name attached to it:

```
patrik <random-uuid@somewhere.com>
```

Currently you cannot copy this address and paste it as a recipient into
the composer on iOS or Android, because they treat the entire string as
if it was an email address and fail, saying "yoo, what's that trash".

This commit extends our UniFFI interface with a new function,
`new_recipient()`, that both clients can use in order to parse the
strings copy-pasted by users before they are passed to other SDK
functions such as `add_single_recipient()`.