fix(ET-5086): Handle draft creation failure due to sync issues
There is a chance that if we have some state update that slips through
our defences we can trigger a constraint invalidation in the database.
This happens when we pull in a draft from the server before we receive
the response from the server.

Safeguard have been added with 64ab7bdf81903d1cd0092ec216b6e7fa2842fa9f
and, just in case, When we detect this we simply delete the previous
message that should not exists.