fix(ET-5139): Handle unplanned conversation id change in draft reply
Under certain circumstances, such as changing the subject of a reply,
the returned message from the server can have a different conversation
id.

This was not handled and could lead to corrupted state from which we can
not recover from until the users performs some actions on the server
that would trigger updates to reset these modifications.

We now catch this after the draft is created and updated and create a
new placeholder conversation. This is fine as the full conversation will
become available after sending. Additionally, since this is a new
conversation, it can only be seen from Drafts and Sent folders, which in
turn only display messages.

Finally, we also remove so noop label applications after the message was
sent.