fix(ET-5222): Fix invalid isProton setting in message contacts
The  `is_proton` field of the `MessageAddress` is meant for official
proton communications. We currently fill in this data from a validity
check which validates if this address is actually powered by our email
service.

We now keep track of both states in the validation response to make sure
the state is correctly propagated.