chore(ET-3605): Revert API-first approach
# Reason

The API-first approach completely skips local resolution of the action.
If you don't have an internet (and even in the case where message is
already present in the Database), you won't ever see a change until
network re-occurs. In the next commit message I will make different
approach to the solution