fix(ET-3605)*: Make sure we are trying to execute actions immediately
This approach (of having an immediate remote part execution) is
necessary to avoid having following issue:

1. Get a notification
2. Move a mail to trash with an action
3. DO NOT open the app
4. Open the app on separate device.

Since the queue is frozen in the background there is no guarantee
that the remote part of the action would be executed in acceptable time.

Therefore it is better for us to trigger it immediately and enqueue
the action only if we fail.