refactor(ET-5183): Rebase move message and conversation
Introduce rebase tests for the move action. Unfortunately to correctly
support this new setup the `ActionMoveData` had to be refactored to
correctly track the relevant information per modified id. It would be
impossible to correctly rebase the changes otherwise.

The move action now keeps track of which modified were modified per
item rather than conversations or messages. This is necessary in order
to determine the rebase changes and to make sure that when we
undo/revert a conversation move, it correctly re-applies to labels to
the messages.

This in turn imposes a requirement that all the messages for the
conversation should be present for the undo/revert. This will be
enforced by the `MailScroller` once INBE-2529 hits production.

Finally the undo move has been simplified into label and mark unread
actions over messages, as that is the only way to make all the state
changes behave correctly.