fix(ET-5274): Do not delete draft body if send is not yet complete
Due to new rules related to fetching new messages from the server it is
possible that a conversation will attempt to fetch the sent message from
the server as the sending of the message is still ongoing.

If we detect the message has still not been sent, we do not allow the
data to be removed.