fix(ET-3605): Resume background tasks for push notification actions
Creating a new user context halts until application enters foreground
because register_subscribers() uses MPSC queue and the Consumer lives
in background task.

The solution is to resume background, do the job and then pause the
background again.

We decided to move `execute_notification_quick_action` to `MailSession`
because:

1. That created user context has to be very short lived (note, we do
not store it in the map!).
2. We must ensure the background tasks are paused asap.