refactor(ET-5183): Rebase LabelAs actions
Update LabelAs actions to support rebase correctly. Similar to the
move to actions, we now record the list of modified messages so we can
correctly revert and determine if we do in fact have changes.

For reasons unknown, the current label as actions do not support undo
for label only operations, this is know possible since it will be know
which messages were modified.