feat(ET-5428): Rebase delete in label
When rebasing this action, we should remark all the messages as deleted
in order to catch new messages which were not yet know. This is what
would happen when the action executes on the server.

When we have backend support we can recalculate the list of messages
based of time stamp intervals aprox.