feat(ET-5429): Rebase Snooze and Unsnooze
No major changes required for snooze, but for unsooze we need to
remember what the new revert time should be.

We also introduced `Converation::snooze_unchecked` as a revert of
unsnooze does not guarantee the previous expiration time will still be
valid. To counter that we issue a resync of the conversation from the
server.