fix(ET-5294): Attachment Metadata vs Message attachments
According to backend, the maximum amount of attachments reported on the
`MessageMetadata` is limited to 10. The full list of attachments is only
available after fetching the body.

This in turn has caused an unfortunate bug in our system where updates
to messages from backend (event/scroller) would wipe all the attachments
associated with the body under certain cases, such as when the body only
contains inline attachments or the body has more than 10 attachments.

To prevent this from happening, we now track the `MessageMetadata`
attachments in a separate table. The `MessageBodyMetadata` retains
ownership over the original table and is no longer affected by this.

Finally, in the migration we also wipe the existing message bodies and
metadata that are not drafts so we can correctly reset this information.