fix(ET-5383): Force resync user data after login
The ticket in question is due to an outdated user model being stored in
the user's device.

It's not entirely clear how this can happen, since this means the model
changed on the server and we never got notified.

The user must have triggered a password change request, which force them
to log in again. At this point, since we have already initialized the
context (in case the db wasn't wiped), so we still have old data.

At some point there should have been an event for the updated user
state, but depending on where the user is when this happens, we can fail
to progress the queue due to previous events.

To fix this, we force the initialization stage in login to re-download
the user data. This is represented as `resync_user` in the
`NewMailUserContextOptions` which replaces the previous initialization
enum.