fix: Legacy feature flags are deserializing
# Context

Even though in the proton API docs [^1] all response fields are marked as
required, in the reality most of them are nullable. This causes our
deserialization logic to fail unexpectedly.

# Solution

I marked expiration and update time as optionals, while global and writable
flags implement #[serde(default)]. We, therefore assume, if the response
is missing `Global` or `Writable` fields, it means it is local and read
only flag.

# Additional improvements

Additionally for the sake of debuggability I added minor traces and
spans.


1: https://protonmail.gitlab-pages.protontech.ch/Slim-API/core/#tag/Feature/operation/get_core-v4-features