fix(ET-5563): Correct Contact Update events
Currently the contact update event is skipping over vcard and contact
email data that is returned in the same event. This in turns causes use
to wipe all the vcards associated with that contact after save. This has
gone unnoticed since we tend to re-fetch this data on demand from the
server.

Since the `Contact` update event also includes the emails, it's not
necessary to apply both contact and contact email events at the same
time.