feat(ET-5182): Move v5 code to new event loop setup
Update the event v5 code to run on the v6 setup in a compatibility mode
so we can later run both v5 and v6 code side by side.

Since in v5 all events are present in the same endpoint, we move the
registration of the events and the subscribers to `mail-common` under
the `MailEventSourceV5` which combines both core and mail events.

The core subscribers are registered with a compatibility wrapper that
translates that forwards this to the respective handler.

Since refresh retries are now handled by the event loop, we had to
replicate them in the `RefershAction`. We also had update the
`join_task` macro to return the real errors that occurred, rather than
anyhow so they can be converted and captured properly. This is required
for the retry logic to kick in.

Finally an `events-v6` feature was added as pre-setup to isolate code
that should not run when this feature is enabled. A follow up MR will
focus on integrating v6 and restructuring the code so we can run the same
logic with both event loops.