refactor*: scroller: Improve terminology: Use CQRS
Scroller communicates with the outside world via two channels, `command`
and `ordered_command`. This terminology is rather confusing, since both
channels are, indeed, ordered - in both cases the semantics are affected
by the order in which messages are sent.

It seems that Paweł intended to implement the CQRS pattern here, so
let's switch the terminology to `query` (aka "unordered command") and
`command` (aka "ordered command").