fix(ET-5603): Fix bug in draft conversation update
There was as piece of code added to handle the possibility that new
conversation created for a new draft was pulled in during the event poll
before the save action finished. Unfortunately this was missing a
constraint that should only run when the existing conversation does not
yet have a remote id.

Since conversation id can change during updates, the previous code would
trigger a deletion of the previous conversation and all of its messages.
If you reply to Conv A and then Change to B and then back to A (e.g.
Subject changes), the last update would wipe ConvA and the parent
message. This would leaving a dangling id for the parent message in the
draft metadata that triggered an invalid foreign key constraint.