fix: Missing system label 13
There is currently a bug in the get labels by id route where system
label 13 is not returned. This is causing some users to fail to apply
events when they receive a conversation or message with this label.

This is being fixed on the backend, but in the mean time we introduce
the label via migration.